Moves import of psycopg2
I moved the import of psycopg2 into the function that uses it. This makes it so that psycopg2 is not a requirement but instead an option. This is a better way to go if a user is not needing the database features.

I also added a graceful way of exiting the application in the event tht the psycopg2 library has not been installed but the database features are used.